나누기 수정

+++ src/day3/Calculator.java
@@ -11,7 +11,7 @@ public class Calculator {
         System.out.println(a*b);
     }
     public void divide(int a,int b){
-        System.out.println(a/b);
+        System.out.println((float) a/b);
     }
